增补设备详情页面的输出模式

+++ app/src/main/java/com/xinrui/smart/fragment/ControlledFragment.java
@@ -290,7 +290,7 @@ public class ControlledFragment extends Fragment{
                         for (int i=0;i<content.length();i++){
                             JSONObject device=content.getJSONObject(i);
                             if (device!=null){
-                                int id=device.getInt("id");
+                                long id=device.getLong("id");
                                 String deviceName=device.getString("deviceName");
                                 int type=device.getInt("type");
                                 int houseId=device.getInt("houseId");
@@ -304,7 +304,7 @@ public class ControlledFragment extends Fragment{
                                 list.add(deviceChild);
                                 if (controlled==1){
                                     controlledDeviceChildren.add(deviceChild);
-                                    contollledDeviceChildMap.put((long)id,deviceChild);
+                                    contollledDeviceChildMap.put(id,deviceChild);
                                 }
                             }
                         }
